Refactor to allow for offers in other sports, account for variable number of games across sports, remove debug print statement.
